feat: add multer middleware! now u POST '/posts/' supports image upload through formData type!
here's the walk through:
u have to assign a middleware in the createBlogPost handler, the middleware is basically a multer middleware that is already configured for like what is the limit size, what file types that it accepts..
if the request's value is valid, multer middleware will append a new key to the request! on the req.file.. but when u access it in the controller, u have to validate it's value cuz it's type is still Express.Multer.File or undefined.
then, we can use sharp to save/write the file data to our filesystem! we have to pass in our image as a buffer, then we can resize the image with sharp, and tell sharp where do we want to save the image in our file system!
then yeah, just save the blogImage url in mongoDB as a Url to our server's Url + the file directory that we save the blog image into!
don't forget to use the express.static middleware so that our server can serve static files!