feat: add authentication with passportjs! it is basically an authentication middleware for our server..
first, we have to create a configuration for our passport! just make it in a different file to organize things, and choose your desired strategy.. here I use local strategy, which is just a username and password login..

so the localStrategy is a class! and it receives a callback that has 3 params, a username, pass, and cb, we can use the cb to determine wether the verification is valid or not. and yeah. we do the credentials verivication here. like checking if the user exists or is already signed in in a social provider, and wether the password match.. then if all good, we can just send the user info to the second argument of the cb! it will attach the user to our request object. and also will send the cookie to the user.

and we also have to configure the serializeUser func to know what is it that we want to save in the session..
also we have to configure the deserializeUser to get the session from our database when we login..

then we also have to configure the sessionConfig for our express-session! like the secret and maxAge.. and where do we want to store the session, well for now it is our mongodb..