feat: add getAuthenticatedUser and logout endpoint!
the getAuthenticatedUser will just send the user's detail by using the req.user object from passport, inside it is the user's id, we can use that to fetch by Id to our db. and rsend the result as a rrepsonse.

the logout func is just using the logout function from passport that is accessible from our requset object.