feat: add social auth with passportJS (google & github login) we create 2 new strategies for passportJS, for google oauth and github oauth!
we check if the user with the corresponding provider's id exist..
if no.. check if we can get the email from the provider
if yes, get the user by email, and update the provider's Id field on the user,
if no user is found by email, create new user with the provider's Id and the email set to the email

if we cannot acccess the email from the provider, just create the user also but but without the email. just the provider's Id

also I made a setSessionReturnTo middleware, to add a returnTo key to the session object on the request object! cuz it will be needed on the passport authenticate successReturnToOrRedirect options..
so we send the returnTo query from the frontEnd when hitting the login by provider endpoint, and we can attach the returnTo query to the session object on the req, the value of the req.session.returnTo would be consumed by the successRetrnToOrRedirect func, and the user would be redirected to that url! cool indeed