feat: add email verification with nodemailer! I use brevo for the email provider!
also created the endpoint that receive the email verification code request, that will validate the email whether the email has been taken or not,
then create a new entry to the emailVerificationCodeTokens documents, which will only live for 10 minutes. (set whe creating the model),
after that, the endpoint will execute the function to send the email via nodeMailer that will send the email with brevo as it's host / email provider!

ALSO, updated the signup in userController, where we will receive the verificationCode and check to the db!