feat: change the place where I store the session, before, express-session is set up to store the session in mongodb with the connect-mongo package that will make a collection for the sessions..
but now, the session is stored in our server locally with redis! we store the sessions by configuring the express-session by changing the store value to use redisStore instance from connect-redis package!
and ofcouse updated the destroyAllActiveSession for user that decides to update their password by submitting the forgot password form so it will delete the sessions inside our redis db instead of checking the sessions collection in mongodb atlas..

well now our session should be lightning fast!
this is important cuz remember, we are using sessions and not JWT!
with authentication using session, on each request, the server has to get the session stored and match it with the request's cookie to check wether it is stored or not.. so the more optimized the session retrieval, the better our server's performance would be!

nice